Small refactoring on JavaScript style for brackets, color conversion for "0x" hex codes, version string updated.
I'll apply the "Google JavaScript Style Guide", first change is the
position of the brackets…
http://google-styleguide.googlecode.com/svn/trunk/javascriptguide.xml

The color code provided in the following format 0xRRGGBB or 0xRGB will
be converted to use "#" instead of "0x".

The version number was outdated! (Used as the plugin identifier).

The compresses version of the script will be updated as well.